requirements.txt
Django==5.1.7
djangorestframework==3.14.0
drf-yasg==1.21.4
django-cors-headers==3.13.0